[gui/Brush 문자 출력 문제 수정] - drawChar() 함수가 경우에 따라 문자가 정상 출력되지 않는 문제 발견 - xoffset이 음수일 경우가 있고 그게 원인으로 보임 - 임시로 관련 문제 수정

+++ src/gui/yss_Brush.cpp
@@ -693,14 +693,14 @@ uint8_t Brush::drawChar(Position_t pos, uint32_t utf8)
 	Font::fontInfo_t *fontInfo = mFont->getFontInfo(utf8);
 	uint8_t *fontFb, color;
 	int32_t  index = 0;
-	uint16_t width = fontInfo->width, height = fontInfo->height, offset = 0, xoffset;
+	int16_t width = fontInfo->width, height = fontInfo->height, offset = 0, xoffset;
 	int16_t xs = pos.x, ys = pos.y + (int8_t)fontInfo->ypos;
 	
 	if(fontInfo == 0)
 		return 0;
 	
 	fontFb = fontInfo->data;
-	xoffset = fontInfo->xpos;
+	xoffset = (int8_t)fontInfo->xpos;
 
 	if(xoffset == 0) // 문자의 앞 여백이 없을 경우
 		xoffset = 1; // 문자의 앞 여백 하나를 추가해줌
